style(nav): Center navbar logo and links together
Updates the navbar layout to group the brand logo and the navigation
links together in the center of the header.

Previously, `justify-content: space-between` pushed the elements to
opposite ends of the navbar. This has been changed to `center` to
create a more unified and focused navigation block. A `gap` property
has also been added to ensure there is adequate spacing between the
logo and the text links.